rename sdxl stub module file
