+++ caikit_computer_vision/modules/text_to_image/sdxl_stub.py
